--- a/binding.js
+++ b/binding.js
@@ -1,0 +1,2 @@
+require.addon = require('require-addon')
+module.exports = require.addon('.', __filename)