### Changes:
- ported `napi` to `libjs` + `bare-compat-napi`
- fixed bug: no more double `close_handle_scope()`
- add asserts to all returns codes
- fix `SN_ASYNC_TASK` memleak
- moved deprecated `Buffer.slice()` calls to `Buffer.subarray()` in `test/*`
- fastcalls introduced:
  - `sodium.crypto_generichash()`
  - `sodium.crypto_generichash_batch()`
  - `sodium.crypto_generichash_init()`
  - `sodium.crypto_generichash_update()`
  - `sodium.crypto_generichash_final()`
  - `sodium.crypto_sign_verify_detached()`
  - `sodium.crypto_box_unseal()`
  - `sodium.crypto_stream_xor()`
  - `crypto_secretstream_xchacha20poly1305_push()`
  - `crypto_secretstream_xchacha20poly1305_pull()`

### Perf-report

#### `node`
| api                          | n-runs       | main (ms)  | libjs (ms) | speedup   |
|------------------------------|--------------|-------|-------|-----------|
| `crypto_generichash`         | 92e6         | 18565 | 18470 | __0.5%__  |
| `crypto_sign_verify_detached` | 2e6          | 90581 | 87565 | __3.4%__  |
| `crypto_box_unseal`          | 4e5          | 24548 | 23791 | __3.2%__  |
| `crypto_generichash_batch`   | {4096} x 1e5 | 36114 | 32263 | __11.9%__ |



#### `bare`

| api                          | n-runs       | main (ms) | libjs (ms) | speedup   |
|------------------------------|--------------|-------|-------|-----------|
| `crypto_generichash`         | 92e6         | 18149 | 15968 | __13.7%__ |
| `crypto_sign_verify_detached` | 2e6          | 90486 | 87711 | __3.2%__  |
| `crypto_box_unseal`          | 4e5          | 24585 | 23530 | __4.5%__  |
| `crypto_generichash_batch`   | {4096} x 1e5 | 40500 | 22233 | __82.2%__ |
| `crypto_stream_xor` | 2e6 | 5565 | 5051 | __10.2%__|
| `crypto_secretstream_xchacha20poly1305_push` | 1e6 | 2220 | 2001 | __10.9%__|
| `crypto_secretstream_xchacha20poly1305_pull` | 1e6 | 2220 | 2001 | __10.9%__|


### Depends
- https://github.com/holepunchto/libjs/pull/12
- https://github.com/holepunchto/bare-compat-napi/pull/1